requirements.txt
torch
torchvision
torchattacks
Pillow
streamlit
face_recognition
numpy